// frontend/types/etl.ts
export interface ETLSummary {
  date: string;
  day: string;
  fx_cnt: number;
  close_cnt: number;
  vol_cnt: number;
  close_eur_cnt: number;
  close_usd_cnt: number;
  vol_eur_cnt: number;
  vol_usd_cnt: number;
  mcap_cnt: number;
  mcap_eur_cnt: number;
  mcap_usd_cnt: number;
  created_at: string;
}

 